% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_connect.R
\name{print.ACSData}
\alias{print.ACSData}
\title{Print method for ACSData}
\usage{
\method{print}{ACSData}(x, ...)
}
\arguments{
\item{x}{An \code{ACSData} object.}

\item{...}{Ignored.}
}
\description{
Print method for ACSData
}
