% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_connect.R
\name{print.DeconvInput}
\alias{print.DeconvInput}
\title{Print method for DeconvInput}
\usage{
\method{print}{DeconvInput}(x, ...)
}
\arguments{
\item{x}{A \code{DeconvInput} object.}

\item{...}{Ignored.}
}
\description{
Print method for DeconvInput
}
