% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_connect.R
\name{summary.ACSData}
\alias{summary.ACSData}
\title{Summary method for ACSData}
\usage{
\method{summary}{ACSData}(object, ...)
}
\arguments{
\item{object}{An \code{ACSData} object.}

\item{...}{Ignored.}
}
\description{
Summary method for ACSData
}
